resume nehancer/requirements.txt
fastapi
uvicorn
openai
python-dotenv
google-generativeai